requirements.txt
Flask==3.0.3
langchain==0.2.16
langchain-ollama==0.1.3
mlflow==3.8.1 #2.17.2
requests==2.32.5
streamlit==1.41.1
python-dotenv